Correctly delete object from the session DB
