+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/DatabaseCleaner.kt
@@ -56,7 +56,7 @@ internal class DatabaseCleaner @Inject constructor(@SessionDatabase private val 
         }
     }
 
-    private suspend fun cleanUp(realm: Realm, threshold: Long) {
+    private fun cleanUp(realm: Realm, threshold: Long) {
         val numberOfEvents = realm.where(EventEntity::class.java).findAll().size
         val numberOfTimelineEvents = realm.where(TimelineEventEntity::class.java).findAll().size
         Timber.v("Number of events in db: $numberOfEvents | Number of timeline events in db: $numberOfTimelineEvents")
@@ -76,20 +76,7 @@ internal class DatabaseCleaner @Inject constructor(@SessionDatabase private val 
                 chunk.numberOfTimelineEvents = chunk.numberOfTimelineEvents - eventsToRemove.size
                 eventsToRemove.forEach {
                     val canDeleteRoot = it.root?.stateKey == null
-                    if (canDeleteRoot) {
-                        it.root?.deleteFromRealm()
-                    }
-                    it.readReceipts?.readReceipts?.deleteAllFromRealm()
-                    it.readReceipts?.deleteFromRealm()
-                    it.annotations?.apply {
-                        editSummary?.deleteFromRealm()
-                        pollResponseSummary?.deleteFromRealm()
-                        referencesSummaryEntity?.deleteFromRealm()
-                        reactionsSummary.deleteAllFromRealm()
-                    }
-                    it.annotations?.deleteFromRealm()
-                    it.readReceipts?.deleteFromRealm()
-                    it.deleteFromRealm()
+                    it.deleteOnCascade(canDeleteRoot)
                 }
                 // We reset the prevToken so we will need to fetch again.
                 chunk.prevToken = null
