+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/EventAnnotationsSummaryEntity.kt
@@ -30,4 +30,12 @@ internal open class EventAnnotationsSummaryEntity(
 ) : RealmObject() {
 
     companion object
+
+    fun deleteOnCascade() {
+        reactionsSummary.deleteAllFromRealm()
+        editSummary?.deleteFromRealm()
+        referencesSummaryEntity?.deleteFromRealm()
+        pollResponseSummary?.deleteFromRealm()
+        deleteFromRealm()
+    }
 }
