+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/ReadReceiptsSummaryEntity.kt
@@ -33,4 +33,9 @@ internal open class ReadReceiptsSummaryEntity(
     val timelineEvent: RealmResults<TimelineEventEntity>? = null
 
     companion object
+
+    fun deleteOnCascade() {
+        readReceipts.deleteAllFromRealm()
+        deleteFromRealm()
+    }
 }
