+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/TimelineEventEntity.kt
@@ -20,6 +20,7 @@ import io.realm.RealmObject
 import io.realm.RealmResults
 import io.realm.annotations.Index
 import io.realm.annotations.LinkingObjects
+import org.matrix.android.sdk.internal.extensions.assertIsManaged
 
 internal open class TimelineEventEntity(var localId: Long = 0,
                                         @Index var eventId: String = "",
@@ -38,4 +39,14 @@ internal open class TimelineEventEntity(var localId: Long = 0,
     val chunk: RealmResults<ChunkEntity>? = null
 
     companion object
+
+    fun deleteOnCascade(canDeleteRoot: Boolean) {
+        assertIsManaged()
+        if(canDeleteRoot) {
+            root?.deleteFromRealm()
+        }
+        annotations?.deleteOnCascade()
+        readReceipts?.deleteOnCascade()
+        deleteFromRealm()
+    }
 }
